Try not to swallow exceptions while subscribing
If the query can't be encoded (or any other checked or unchecked Exception
occurs), the async SubscribeClientOperation will swallow the exception,
leaving no way (even a log message) to see why it failed.

This creates a new LiveQueryException subclass (UnknownException) to
represent an unexpected RuntimeException that can then be handled by the
HandleErrorCallback#onError() callback.