Fix #21: connect and disconnect on the background executor (#22)
* Fix #21: connect and disconnect on the background executor

TubeSock already uses a background thread for `open()`, but it keeps the
`close()` call on the calling thread. So `open()` is safe on the main thread,
but `close()` is not.

* Add test for #21

* Update the tests to always use a PauseableExecutor

The end result will be the same for all the other tests, since it will only
queue up tasks while it is paused, and it is unpaused by default.

To control when the tasks run, the test should first pause the executor
before the tasks get enqueued.