Clarify the logic of initiating a connection during subscribe()
This also creates a new connectIfNeeded() method, which can be used to
initiate a connection if not already connected or connecting.

The original intent was to connectIfNeeded() during subscribe. While some may
argue that's not necessary, and the client should instead be explicit about
what is intended, this change preserves that original intention.

As it stands, the only way to queue up pending subscriptions _without_
initiating a connection would be to call disconnect() first, even though the
client hasn't been connected yet. That would set the user-initiated flag,
which would therefore refuse to auto-connect during subscribe().